Crate aster_util

The util of Asterinas.

Modules

coeff

This module provides an abstraction Coeff to server for efficient and accurate calculation of fraction multiplication.

dup

fixed_point

A lightweight fixed-point number implementation optimized for kernel use.

mem_obj_slice

Provides Slice for quick duplication and slicing over memory objects, such as UFrame, USegment, IoMem, DmaStream, etc.

per_cpu_counter

A fast and scalable per-CPU counter.

printer

ranged_integer

Ranged integer types.

safe_ptr

slot_vec

Macros

field_ptr

Create a safe pointer for the field of a struct.