aster_util::mem_obj_slice

Struct Slice

pub struct Slice<MemObj> { /* private fields */ }

A slice of a memory object.

Implementations

impl<MemObj: HasSize> Slice<MemObj>

pub fn new(mem_obj: MemObj, offset: Range<usize>) -> Self

Creates a new slice from the given offset range.

The range is relative to the start of the underlying memory object.

Panics

The function panics if

the range is empty or negative, or

the range is not within the range of the underlying memory object.

pub fn mem_obj(&self) -> &MemObj

Returns the underlying memory object of this slice.

pub fn offset(&self) -> &Range<usize>

Returns the offset range of this slice within the underlying memory object.

impl<MemObj: HasSize + Clone> Slice<MemObj>

pub fn slice(&self, range: Range<usize>) -> Self

Returns a slice of this memory object slice.

The range is relative to the start of the memory object slice, not the start of the underlying memory object.

Panics

The function panics if

the range is empty or negative, or

the range is not within the range of this memory object slice.

impl<MemObj: HasSize + Borrow<Arc<DmaStream>>> Slice<MemObj>

pub fn sync_from_device(&self) -> Result<()>

Synchronizes the slice of streaming DMA mapping from the device.

The method will call DmaStream::sync_from_device with the offset range of this slice.

pub fn sync_to_device(&self) -> Result<()>

Synchronizes the slice of streaming DMA mapping to the device.

The method will call DmaStream::sync_to_device with the offset range of this slice.