aster_util::printer

Struct VmPrinter

pub struct VmPrinter<'a, 'b> { /* private fields */ }

A specialized printer for formatted text output.

VmPrinter is designed to handle the common pattern in kernel where kernel code needs to generate formatted text output (like status information, statistics, or configuration data) and write it to user space with proper offset handling.

Examples

use aster_util::printer::VmPrinter;
use ostd::{mm::VmWriter, Pod};

let mut buf = [0u8; 3];
let mut writer = VmWriter::from(buf.as_bytes_mut()).to_fallible();
let mut printer = VmPrinter::new_skip(&mut writer, 3);

let res = writeln!(printer, "val: {}", 123);
assert!(res.is_ok());

assert_eq!(printer.bytes_written(), 3);
assert_eq!(&buf, b": 1");

Implementations

impl<'a, 'b> VmPrinter<'a, 'b>

pub fn new_skip(writer: &'a mut VmWriter<'b>, bytes_to_skip: usize) -> Self

Creates a new VmPrinter that skips the first bytes_to_skip bytes and prints the remaining bytes to writer.

pub fn bytes_written(&self) -> usize

Returns the total number of bytes written to the underlying writer.

pub fn write_fmt(&mut self, args: Arguments<'_>) -> Result<(), VmPrinterError>

Writes formatted content to the underlying writer.