aster_util::safe_ptr

Struct SafePtr

pub struct SafePtr<T, M, R = Full> { /* private fields */ }

Safe pointers.

Overview

Safe pointers allows using pointers to access memory without unsafe code, which is a key enabler for writing device drivers in safe Rust.

To ensure its soundness, safe pointers (SafePtr<T, M, _>) have to be more restricted than raw pointers (*const T or *mut T). More specifically, there are three major restrictions.

A safe pointer can only refer to a value of a POD type T: Pod, while raw pointers can do to a value of any type T.

A safe pointer can only refer to an address within a virtual memory object of type M: VmIo (e.g., VMAR and VMO), while raw pointers can do to an address within any virtual memory space.

A safe pointer only allows one to copy values to/from the target address, while a raw pointer allows one to borrow an immutable or mutable reference to the target address.

The expressiveness of safe pointers, although being less than that of raw pointers, is sufficient for our purpose of writing an OS kernel in safe Rust.

In addition, safe pointers SafePtr<T, M, R> are associated with access rights, which are encoded statically with type R: TRights.

Examples

Constructing a safe pointer

An instance of SafePtr can be created with a VM object and an address within the VM object.

let u32_ptr: SafePtr<u32, Vec<u8>, _> = {
    let vm_obj = VmoOptions::new(PAGE_SIZE).alloc().unwrap();
    let addr = 16;
    SafePtr::new(vm_obj, addr)
};

The generic parameter M of SafePtr<_, M, _> must implement the VmIo trait. The most important VmIo types are Vmar, Vmo, IoMem, and UFrame. The blanket implementations of VmIo also include pointer-like types that refer to a VmIo type. Some examples are &Vmo, Box<Vmar>, and Arc<IoMem>.

The safe pointer itself does not and cannot guarantee that its address is valid. This is because different VM objects may interpret addresses differently and each VM object can have different restrictions for valid addresses. So the detection of invalid addresses is delayed to the time when the pointers are actually read from or written to.

Initially, a newly-created safe pointer has all access rights.

Reading and writing a safe pointer

The value pointed to by a safe pointer can be read or written with the read or write method. Both methods may return errors. The possible reasons of error are determined by the underlying VM objects.

u32_ptr.write(1234).unwrap();
assert!(u32_ptr.read().unwrap() == 1234);

Manipulating a safe pointer

The address of a safe pointer can be obtained by the addr method. The address can be updated by assigning a new value with the set_addr method or updated incrementally through methods like add, offset, byte_addr, byte_offset.

The VM object of a safe pointer can also be obtained or updated through the vm and set_vm methods. A new safe pointer that is backed by the same VM object of an existing safe pointer can be obtained through the borrow_vm method.

As an example, the code below shows how the add and borrow_vm methods can be used together to to iterate all values pointed to by an array pointer.

fn collect_values<T>(array_ptr: &SafePtr<T, M, _>, array_len: usize) -> Vec<T> {
    let mut curr_ptr: SafePtr<T, &M, _> = array_ptr.borrow_vm();
    (0..array_len)
        .iter()
        .map(|_| {
            let val = curr_ptr.read().unwrap();
            curr_ptr.add(1);
            val
        })
        .collect()
}

The data type of a safe pointer can be converted with the cast method.

let u8_ptr: SafePtr<u8, _, _> = u32_ptr.cast();

Reading and writing the fields of a struct

Given a safe pointer that points to a struct (say, Foo), one can read the value of its field as follows.

pub struct Foo {
    first: u64,
    second: u32,
}

fn read_second_field<M: VmIo>(ptr: &SafePtr<Foo, M, _>) -> u32 {
    let field_ptr = ptr
        .borrow_vm()
        .byte_add(offset_of!(Foo, second) as usize)
        .cast::<u32>();
    field_ptr.read().unwrap()
}

But this coding pattern is too tedius for such a common task. To make the life of users easier, we provide a convenient macro named field_ptr, which can be used to obtain the safe pointer of a field from that of its containing struct.

fn read_second_field<M: VmIo>(ptr: &SafePtr<Foo, M, _>) -> u32 {
    let field_ptr = field_ptr!(ptr, Foo, second);
    field_ptr.read().unwrap()
}

Access rights

A safe pointer may have a combination of three access rights: Read, Write, and Dup.

Implementations

impl<T, M> SafePtr<T, M>

pub fn new(vm_obj: M, offset: usize) -> Self

Create a new instance.

Access rights

The default access rights of a new instance are Read, Write, and Dup.

impl<T: Pod, M: VmIo, R: TRights> SafePtr<T, M, TRightSet<R>>

pub fn read(&self) -> Result<T>
where R: SetContain<Read>, SetContainOp<R, Read>: IsTrue,

Reads the value from the pointer.

Access rights

This method requires the Read right.

pub fn read_slice(&self, slice: &mut [T]) -> Result<()>
where R: SetContain<Read>, SetContainOp<R, Read>: IsTrue,

Reads a slice of values from the pointer.

Access rights

This method requires the Read right.

pub fn write(&self, val: &T) -> Result<()>
where R: SetContain<Write>, SetContainOp<R, Write>: IsTrue,

Overwrites the value at the pointer.

Access rights

This method requires the Write right.

pub fn write_slice(&self, slice: &[T]) -> Result<()>
where R: SetContain<Write>, SetContainOp<R, Write>: IsTrue,

Overwrites a slice of values at the pointer.

Access rights

This method requires the Write right.

pub fn copy_from<M1: HasVmReaderWriter, R1>( &self, ptr: &SafePtr<T, M1, TRightSet<R1>>, ) -> Result<()>
where R: SetContain<Write>, SetContainOp<R, Write>: IsTrue, R1: SetContain<Read> + TRights, SetContainOp<R1, Read>: IsTrue,

Copies the value from another pointer.

Errors

This method will fail if

either pointerโs underlying VM object is too short to hold T or

the memory copy between two pointers fails.

Access rights

This method requires

the Write right for this pointer and

the Read right for another pointer.

impl<T: PodOnce, M: VmIoOnce, R: TRights> SafePtr<T, M, TRightSet<R>>

pub fn read_once(&self) -> Result<T>
where R: SetContain<Read>, SetContainOp<R, Read>: IsTrue,

Reads the value from the pointer using one non-tearing instruction.

Access rights

This method requires the Read right.

pub fn write_once(&self, val: &T) -> Result<()>
where R: SetContain<Write>, SetContainOp<R, Write>: IsTrue,

Overwrites the value at the pointer using one non-tearing instruction.

Access rights

This method requires the Write right.

impl<T, M, R> SafePtr<T, M, R>

pub const fn is_aligned(&self) -> bool

pub fn add(&mut self, count: usize)

Increase the address in units of bytes occupied by the generic T.

pub fn offset(&mut self, count: isize)

Increase or decrease the address in units of bytes occupied by the generic T.

pub fn byte_add(&mut self, bytes: usize)

Increase the address in units of bytes.

pub fn byte_offset(&mut self, bytes: isize)

Increase or decrease the address in units of bytes.

impl<T, M, R> SafePtr<T, M, R>

pub const fn vm(&self) -> &M

pub fn set_vm(&mut self, vm_obj: M)

impl<T, M, R: Clone> SafePtr<T, M, R>

pub fn borrow_vm(&self) -> SafePtr<T, &M, R>

Construct a new SafePtr which will point to the same address

impl<T, M, R> SafePtr<T, M, R>

pub fn cast<U>(self) -> SafePtr<U, M, R>

Cast the accessed structure into a new one, which is usually used when accessing a field in a structure.

impl<T, M, R: TRights> SafePtr<T, M, TRightSet<R>>

pub fn restrict<R1>(self) -> SafePtr<T, M, TRightSet<R1>>
where R1: Set + TRights, R: SetInclude<R1>, SetIncludeOp<R, R1>: IsTrue,

Construct a new SafePtr and restrict the rights of it.

Access rights

This method requires the target rights to be a subset of the current rights.

impl<T, R, D: DmaDirection> SafePtr<T, DmaStream<D>, R>

pub fn sync_from_device(&self) -> Result<()>

Synchronizes the slice of streaming DMA mapping from the device.

The method will call ostd::mm::dma::DmaStream::sync_from_device with the offset range of this pointer.

pub fn sync_to_device(&self) -> Result<()>

Synchronizes the slice of streaming DMA mapping to the device.

The method will call ostd::mm::dma::DmaStream::sync_to_device with the offset range of this pointer.