aster_util::slot_vec

Struct SlotVec

pub struct SlotVec<T> { /* private fields */ }

SlotVec is the variant of Vector. It guarantees that the index of one item remains unchanged during adding or deleting other items of the vector.

Implementations

impl<T> SlotVec<T>

pub const fn new() -> Self

New an empty vector.

pub fn with_capacity(capacity: usize) -> Self

pub fn is_empty(&self) -> bool

Return true if the vector contains no items.

pub fn len(&self) -> usize

Return the number of items.

pub fn slots_len(&self) -> usize

Return the number of slots.

pub fn get(&self, idx: usize) -> Option<&T>

Get the item at position idx.

Return None if idx is out of bounds or the item is not exist.

pub fn get_mut(&mut self, idx: usize) -> Option<&mut T>

Get the mutable reference of the item at position idx.

Return None if idx is out of bounds or the item is not exist.

pub fn put(&mut self, entry: T) -> usize

Put an item into the vector. It may be put into any existing empty slots or the back of the vector.

Return the index of the inserted item.

pub fn put_at(&mut self, idx: usize, item: T) -> Option<T>

Put and return the item at position idx.

Return None if the item is not exist.

pub fn remove(&mut self, idx: usize) -> Option<T>

Remove and return the item at position idx.

Return None if idx is out of bounds or the item has been removed.

pub fn idxes_and_items(&self) -> impl Iterator<Item = (usize, &T)>

Create an iterator which gives both of the index and the item. The index may not be continuous.

pub fn iter(&self) -> impl Iterator<Item = &T>

Create an iterator which just gives the item.